fix(arch): satisfy checker boundary guards for deferred mirror replay
- extract def_mapping.rs tests into def_mapping_tests.rs (2096 -> 1867 lines)

- add deferred_flow_env_writes to checker_context_lifetimes.toml (FileLocalReset/RelationSessionState)

- bump CheckerContext field cap 244 -> 245 and note it in ROADMAP.md
